Refactor stat utilities into focused modules
